[272] * Update to MaNGOS revision 5343+
* Update to MaNGOS revision 5343+
* Remove: Use of strings in InstanceData. Replaced with uint32 identifier. 
* Added: Kael'Thas encounter. 

git-svn-id: https://scriptdev2.svn.sourceforge.net/svnroot/scriptdev2@272 5f9c896b-1e26-0410-94da-f77f675e2462

+++ scripts/zone/karazhan/boss_curator.cpp
@@ -190,7 +190,7 @@ struct MANGOS_DLL_DECL boss_curatorAI : public ScriptedAI
                 }
 
                 //Reduce Mana by 10%
-                int32 mana = (0.1*(m_creature->GetMaxPower(POWER_MANA)));               
+                int32 mana = (int32)(0.1*(m_creature->GetMaxPower(POWER_MANA)));               
                 m_creature->ModifyPower(POWER_MANA, -mana);
                 switch(rand()%4)
                 {
